main.py
from src.extract_data import extract_weather_data
from src.transform_data import data_transformations 
from src.load_data import load_weather_data

import os
from pathlib import Path
from dotenv import load_dotenv

import logging
logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')

env_path = Path(__file__).resolve().parent.parent / 'config' / '.env'
load_dotenv(env_path)

API_KEY = os.getenv('API_KEY')

url = f'https://api.openweathermap.org/data/2.5/weather?q=Votuporanga,SP&units=metric&appid={API_KEY}'
table_name = 'votu_weather'

def pipeline():
    try:
        logging.info("Iniciando pipeline de ETL...")
        extract_weather_data(url)

        logging.info("Extração de dados concluída. Iniciando transformação...")
        df = data_transformations()

        logging.info("Transformação de dados concluída. Iniciando carga...")
        load_weather_data(table_name, df)

        print("\n" + "="*60)
        print("Pipeline de ETL concluído com sucesso!")
        print("="*60)

    except Exception as e:
        logging.error(f"Erro na pipeline: {e}")
        import traceback
        traceback.print_exc()

pipeline()